+++ genium/src/test/resources/smoke/listeners/output/android/com/example/smoke/InternalListener.java
@@ -1,0 +1,8 @@
+/*
+ *
+ * Automatically generated. Do not modify. Your changes will be lost.
+ */
+package com.example.smoke;
+interface InternalListener {
+    void onEvent();
+}