+++ genium/src/test/resources/smoke/listeners/output/android/jni/ProxyConversion.h
@@ -2,6 +2,7 @@
 #include "com_example_smoke_ListenerWithAttributesImplCppProxy.h"
 #include "com_example_smoke_CalculatorListenerImplCppProxy.h"
 #include "com_example_smoke_CalculationResultImplCppProxy.h"
+#include "com_example_smoke_InternalListenerImplCppProxy.h"
 #include "com_example_smoke_ListenersWithReturnValuesImplCppProxy.h"
 template<class T>
 inline void createCppProxy(JNIEnv* env, const ::genium::jni::JniReference<jobject>& obj, ::std::shared_ptr<T>& result) {
@@ -17,6 +18,10 @@ inline void createCppProxy(JNIEnv* env, const ::genium::jni::JniReference<jobjec
 template<>
 inline void createCppProxy(JNIEnv* env, const ::genium::jni::JniReference<jobject>& obj, ::std::shared_ptr<::smoke::CalculationResult>& result) {
     ::genium::jni::CppProxyBase::createProxy<::smoke::CalculationResult, ::smoke::CalculationResultCppProxy>(env, obj, result);
+}
+template<>
+inline void createCppProxy(JNIEnv* env, const ::genium::jni::JniReference<jobject>& obj, ::std::shared_ptr<::smoke::InternalListener>& result) {
+    ::genium::jni::CppProxyBase::createProxy<::smoke::InternalListener, ::smoke::InternalListenerCppProxy>(env, obj, result);
 }
 template<>
 inline void createCppProxy(JNIEnv* env, const ::genium::jni::JniReference<jobject>& obj, ::std::shared_ptr<::smoke::ListenersWithReturnValues>& result) {